1)舰队接口封装完成 2)入侵信息接口封装完成 3)国服ESI地址更换 4)增加相关ESI接口的test

+++ src/main/java/xyz/foolcat/eve/evehelper/domain/system/Observer.java
@@ -33,7 +33,7 @@ public class Observer implements Serializable {
 
     @TableField(value = "croporation_id")
     @Schema(description = "")
-    private Long croporationId;
+    private Integer croporationId;
 
     private static final long serialVersionUID = 1L;
