+++ src/main/java/xyz/foolcat/eve/evehelper/esi/api/CalendarApi.java
@@ -53,7 +53,7 @@ public class CalendarApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-日历事件列表")
-    public Flux<CalendarResponse> queryCharactersCalendar(Long characterId, String datasource, Integer fromEventId, String accessesToken) {
+    public Flux<CalendarResponse> queryCharactersCalendar(Integer characterId, String datasource, Integer fromEventId, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/calendar/?datasource={datasource}&from_event={fromEventId}", characterId, datasource, fromEventId)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -80,7 +80,7 @@ public class CalendarApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-事件详情")
-    public Mono<CalendarEventResponse> queryCharactersCalendarEventId(Long characterId, String datasource, Integer eventId, String accessesToken) {
+    public Mono<CalendarEventResponse> queryCharactersCalendarEventId(Integer characterId, String datasource, Integer eventId, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/calendar/{event_id}/?datasource={datasource}", characterId, eventId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -110,7 +110,7 @@ public class CalendarApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-回复事件")
-    public Mono<Object> updateCharactersCalendarEventId(Long characterId, String datasource, Integer eventId, CalendarEventEnum calendarEventEnum, String accessesToken) {
+    public Mono<Object> updateCharactersCalendarEventId(Integer characterId, String datasource, Integer eventId, CalendarEventEnum calendarEventEnum, String accessesToken) {
         Map<String, String> responseReply = new HashMap<>();
         responseReply.put("response", calendarEventEnum.getValue());
         return apiClient.put().uri("/characters/{character_id}/calendar/{event_id}/?datasource={datasource}", characterId, eventId, datasource)
@@ -142,7 +142,7 @@ public class CalendarApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-事件参与列表")
-    public Flux<CalendarEventAttendeesResponse> queryCharactersCalendarEventIdAttendees(Long characterId, String datasource, Integer eventId, String accessesToken) {
+    public Flux<CalendarEventAttendeesResponse> queryCharactersCalendarEventIdAttendees(Integer characterId, String datasource, Integer eventId, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/calendar/{event_id}/attendees/?datasource={datasource}", characterId, eventId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
