+++ src/main/java/xyz/foolcat/eve/evehelper/esi/api/CharacterApi.java
@@ -45,7 +45,7 @@ public class CharacterApi {
             @Parameter(name = "datasource", description = "服务器数据源", required = true),
     })
     @Operation(summary = "ESI-人物详细公开信息")
-    public Mono<CharacterPublicInfoResponse> queryCharacter(Long characterId, String datasource) {
+    public Mono<CharacterPublicInfoResponse> queryCharacter(Integer characterId, String datasource) {
         return apiClient.get().uri("/characters/{character_id}/?datasource={datasource}", characterId, datasource)
                 .retrieve()
                 .onStatus(HttpStatus::is4xxClientError, response ->
@@ -69,7 +69,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物的代理研究信息列表")
-    public Flux<AgentsResearchResponse> queryCharacterAgentsResearch(Long characterId, String datasource, String accessesToken) {
+    public Flux<AgentsResearchResponse> queryCharacterAgentsResearch(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/agents_research/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -95,7 +95,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物蓝图清单")
-    public Flux<BlueprintResponse> queryCharacterBlueprint(Long characterId, String datasource, Integer page, String accessesToken) {
+    public Flux<BlueprintResponse> queryCharacterBlueprint(Integer characterId, String datasource, Integer page, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/blueprints/?datasource={datasource}&page={page}", characterId, datasource, page)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -118,7 +118,7 @@ public class CharacterApi {
             @Parameter(name = "datasource", description = "服务器数据源", required = true),
     })
     @Operation(summary = "ESI-人物雇佣记录")
-    public Flux<CorporationHistoryResponse> queryCharacterCorporationHistory(Long characterId, String datasource) {
+    public Flux<CorporationHistoryResponse> queryCharacterCorporationHistory(Integer characterId, String datasource) {
         return apiClient.get().uri("/characters/{character_id}/corporationhistory/?datasource={datasource}", characterId, datasource)
                 .retrieve()
                 .onStatus(HttpStatus::is4xxClientError, response ->
@@ -144,7 +144,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-CSPA花费")
-    public Mono<Float> queryCharacterCspa(Long characterId, String datasource, List<Integer> characters, String accessesToken) {
+    public Mono<Float> queryCharacterCspa(Integer characterId, String datasource, List<Integer> characters, String accessesToken) {
         return apiClient.post().uri("/characters/{character_id}/cspa/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .contentType(MediaType.APPLICATION_JSON)
@@ -171,7 +171,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-跳跃疲劳时间")
-    public Mono<FatigueResponse> queryCharacterFatigue(Long characterId, String datasource, String accessesToken) {
+    public Mono<FatigueResponse> queryCharacterFatigue(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/fatigue/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -196,7 +196,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物勋章信息")
-    public Flux<MedalResponse> queryCharacterMedals(Long characterId, String datasource, String accessesToken) {
+    public Flux<MedalResponse> queryCharacterMedals(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/medals/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -221,7 +221,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物通知消息")
-    public Flux<NotificationResponse> queryCharacterNotification(Long characterId, String datasource, String accessesToken) {
+    public Flux<NotificationResponse> queryCharacterNotification(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/notifications/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -246,7 +246,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-添加到联系人通知")
-    public Flux<NotificationContactResponse> queryCharacterNotificationContact(Long characterId, String datasource, String accessesToken) {
+    public Flux<NotificationContactResponse> queryCharacterNotificationContact(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/notifications/contacts/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -269,7 +269,7 @@ public class CharacterApi {
             @Parameter(name = "datasource", description = "服务器数据源", required = true),
     })
     @Operation(summary = "ESI-人物肖像图标地址")
-    public Mono<IconResponse> queryCharacterPortrait(Long characterId, String datasource) {
+    public Mono<IconResponse> queryCharacterPortrait(Integer characterId, String datasource) {
         return apiClient.get().uri("/characters/{character_id}/portrait/?datasource={datasource}", characterId, datasource)
                 .retrieve()
                 .onStatus(HttpStatus::is4xxClientError, response ->
@@ -293,7 +293,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物角色列表")
-    public Mono<RoleResponse> queryCharacterRoles(Long characterId, String datasource, String accessesToken) {
+    public Mono<RoleResponse> queryCharacterRoles(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/roles/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -318,7 +318,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物声望列表")
-    public Flux<StandingResponse> queryCharacterStanding(Long characterId, String datasource, String accessesToken) {
+    public Flux<StandingResponse> queryCharacterStanding(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/standings/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -344,7 +344,7 @@ public class CharacterApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物职位列表")
-    public Flux<TitleResponse> queryCharacterTitle(Long characterId, String datasource, String accessesToken) {
+    public Flux<TitleResponse> queryCharacterTitle(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/titles/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
