+++ src/main/java/xyz/foolcat/eve/evehelper/esi/api/ContactApi.java
@@ -99,7 +99,7 @@ public class ContactApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物删除人物联系人")
-    public Mono<Object> deleteCharactersContacts(Long characterId, String datasource, List<Long> contactIds, String accessesToken) {
+    public Mono<Object> deleteCharactersContacts(Integer characterId, String datasource, List<Long> contactIds, String accessesToken) {
         return apiClient.delete().uri("/characters/{character_id}/contacts/?datasource={datasource}&contact_ids={contact_ids}", characterId, datasource, contactIds.toArray())
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -124,7 +124,7 @@ public class ContactApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物获取联系人")
-    public Flux<ContactResponse> queryCharactersContacts(Long characterId, String datasource, String accessesToken) {
+    public Flux<ContactResponse> queryCharactersContacts(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/contacts/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -157,7 +157,7 @@ public class ContactApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物添加联系人")
-    public Flux<Long> addCharactersContacts(Long characterId, String datasource, List<Long> contactIds, List<Long> labelIds, Integer standing, boolean watched, String accessesToken) {
+    public Flux<Long> addCharactersContacts(Integer characterId, String datasource, List<Long> contactIds, List<Long> labelIds, Integer standing, boolean watched, String accessesToken) {
         return apiClient.post().uri("/characters/{character_id}/contacts/?datasource={datasource}&label_ids={label_ids}&standing={standing}&watched={watched}", characterId, datasource, labelIds.toArray(), standing, watched)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .body(Mono.just(contactIds), List.class)
@@ -191,7 +191,7 @@ public class ContactApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物更新联系人")
-    public Mono<Object> updateCharactersContacts(Long characterId, String datasource, List<Long> contactIds, List<Long> labelIds, Integer standing, boolean watched, String accessesToken) {
+    public Mono<Object> updateCharactersContacts(Integer characterId, String datasource, List<Long> contactIds, List<Long> labelIds, Integer standing, boolean watched, String accessesToken) {
         return apiClient.put().uri("/characters/{character_id}/contacts/?datasource={datasource}&label_ids={label_ids}&standing={standing}&watched={watched}", characterId, datasource, labelIds.toArray(), standing, watched)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .body(Mono.just(contactIds), List.class)
@@ -217,7 +217,7 @@ public class ContactApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-人物联系人自定义标签")
-    public Flux<ContactLabelResponse> queryCharactersContactsLabel(Long characterId, String datasource, String accessesToken) {
+    public Flux<ContactLabelResponse> queryCharactersContactsLabel(Integer characterId, String datasource, String accessesToken) {
         return apiClient.get().uri("/characters/{character_id}/contacts/labels/?datasource={datasource}", characterId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -242,7 +242,7 @@ public class ContactApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-军团联系人")
-    public Flux<ContactResponse> queryCorporationsContacts(Long corporationId, String datasource, String accessesToken) {
+    public Flux<ContactResponse> queryCorporationsContacts(Integer corporationId, String datasource, String accessesToken) {
         return apiClient.get().uri("/corporations/{corporation_id}/contacts/?datasource={datasource}", corporationId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
@@ -267,7 +267,7 @@ public class ContactApi {
             @Parameter(name = "accessesToken", description = "授权Token", required = true),
     })
     @Operation(summary = "ESI-军团联系人自定义标签")
-    public Flux<ContactLabelResponse> queryCorporationsContactsLabel(Long corporationId, String datasource, String accessesToken) {
+    public Flux<ContactLabelResponse> queryCorporationsContactsLabel(Integer corporationId, String datasource, String accessesToken) {
         return apiClient.get().uri("/corporations/{corporation_id}/contacts/labels/?datasource={datasource}", corporationId, datasource)
                 .header(HttpHeaders.AUTHORIZATION, accessesToken)
                 .retrieve()
